NDC_HID_BE/apps/employee/urls.py
from rest_framework.routers import DefaultRouter
from django.urls import path
from .views import *

router = DefaultRouter()
router.register(r'', EmployeeViewSet, basename='employee')

urlpatterns = [
    path('test/', Test.as_view()),
    *router.urls,
] 