NDC_HID_BE/requirement.txt
amqp==5.3.1
ansicon==1.89.0
arrow==1.3.0
asgiref==3.8.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
blessed==1.20.0
celery==5.4.0
cffi==1.17.1
channels==4.2.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
clr_loader==0.2.7.post0
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
Django==5.1.5
django-cors-headers==4.6.0
django-environ==0.12.0
django-picklefield==3.2
djangorestframework==3.15.2
flower==2.0.1
gevent==24.11.1
greenlet==3.1.1
h11==0.14.0
h2==4.1.0
hpack==4.1.0
humanize==4.11.0
Hypercorn==0.17.3
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jinxed==1.3.0
kombu==5.4.2
priority==2.0.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
psutil==6.1.1
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
pythonnet==3.0.5
pytz==2024.2
redis==3.5.3
service-identity==24.2.0
setuptools==75.8.0
six==1.17.0
sqlparse==0.5.3
tornado==6.4.2
Twisted==24.11.0
txaio==23.1.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
vine==5.1.0
wcwidth==0.2.13
wsproto==1.2.0
zope.event==5.0
zope.interface==7.2
